+++ MikaVM/MikaArchive.h
@@ -29,10 +29,10 @@ public:
 template <typename T>
 MikaArchive& operator<<(MikaArchive& ar, std::vector<T>& vec)
 {
-	int size = vec.size();
+	size_t size = vec.size();
 	ar << size;
 	vec.resize(size);
-	for (int i = 0; i < size; ++i)
+	for (size_t i = 0; i < size; ++i)
 	{
 		ar << vec[i];
 	}
